Implement modular architecture with Google Maps, theme, and config management
